% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{accurracy}
\alias{accurracy}
\title{Calculate Acurracy of the model}
\usage{
accurracy(new_Y, p, verb = FALSE)
}
\arguments{
\item{p}{Vector of fitted probabilities for each case}

\item{verb}{verbose, if True, prints aditional information (logical)}
}
\value{
The acurracy of the model, given the fitted probabilities and new data
}
\description{
Given a set of true values and their corresponding fitted probabilities,
the function calculates the acurracy of the model defined by:
 $1- ,#wrong prediction/# of observations$
}
\examples{
(new_Y_ata, fitted_probs_for_data)
}
