% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpwpm_gibbs.R
\name{bpwpm_gibbs}
\alias{bpwpm_gibbs}
\title{Bayesian Piecewise Polinomial Model}
\usage{
bpwpm_gibbs(Y, X, M, J, K, precision_w = 1, draws = 10^3, tau = NULL,
  w_init = NULL, mu_w_0 = NULL, sigma_w_0_inv = NULL,
  indep_terms = FALSE, verb = FALSE, debug = FALSE)
}
\arguments{
\item{Y}{Response vector of n binary observatios (integers 0,1 - vector of
size n) Can be encoded as a factor a numeric vector.}

\item{X}{Design matrix of n observations and d covariables (numeric - n*d)}

\item{M}{M minus 1 is the degree of the polinomial (integer - M > 0)}

\item{J}{Number of intervals in each dimention (integer - J > 1)}

\item{K}{Order of continuity in the derivatives (integrer - K < M)}

\item{precision_w}{If using the default sigmas for w, a diagonal matrix will
be used. Precision controls its magnitude (numeric - precision > 0)}

\item{draws}{Númber of samples to draw from the Gibbs Sampler (integer - draw
> 0)}

\item{tau}{the initial position of nodes selected by the user. although·
arbitraty they need to match the dimentions. (numeric - (J-1)*d)
  of size d)
  beta (numeric - matrix of size d*d)}

\item{w_init}{Inital value for the Gibbs Sampler Chain (numeric - matrix of
size N*d)}

\item{mu_w_0}{Prior Mean of Beta (numeric - matrix of size N*d)}

\item{sigma_w_0_inv}{sigma_w_0_inv:= Prior Inverse if the Variance-Covariance
Matrices of w (list - d elements, each element is a numeric matrix of size
N*N)}

\item{indep_terms}{Keeps the independent terms in the PWP expansion.
Leaving those terms might lead to identificability errors. (logical)}

\item{verb}{short for verbose, if TRUE, prints aditional information (logical)}

\item{debug}{If TRUE, print even more info to help with debugging (logical)}
}
\value{
An object of the class "bpwpm" containing at the following
components:
\describe{
\item{betas: }{A data frame containing the Gibbs sampler simulation for beta}
\item{w: }{A list of d elements. Each one is a data frame containign the
simulation of the w_j parameters for each dimetnion j.}
\item{Phi: }{The PWP Expansion for input matrix X and nodes selected on
percentiles}
\item{tau: }{Nodes used for training}
\item{M: }{Initial parameters}
\item{J: }{Initial parameters}
\item{K: }{Initial parameters}
\item{d: }{Number of dimentions}
\item{indep_terms: }{Logical. If independent terms are keept}
\item{info}{A string that prints the basic information of the mode. Used for
the summary function.}
}
}
\description{
Gibbs Sampler for simulating draws of parameters for the Bayesian Piece Wise
Polinomial model described on the thesis.
}
\examples{
See the main document of the thesis for a couple of full examples
with its corresponding analysis.
}
