% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{calculate_F}
\alias{calculate_F}
\title{F matrix calculation}
\usage{
calculate_F(Phi, w, d, intercept)
}
\arguments{
\item{Phi}{Piecewise Polinomail expansion for an input matrix X previosly
calculated by \code{\link{calculate_Phi}}}

\item{w}{Set of weights for which to calculate F. Numerical matrix of size
(N*d)}

\item{d}{Number of dimentions, this parameter helps to improve efficiency}

\item{intercept}{Intercept optional paramter. Logical}
}
\value{
F matrix
}
\description{
Private function for calculating the F matrix, described on the thesis.
This is the transformed input matrix that depends on the piecewise polinomial
expansion Phi and a set of weights w.
}
