% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Phi.R
\name{calculate_Phi}
\alias{calculate_Phi}
\title{Piece wise polinomial expansion for X (PWP)}
\usage{
calculate_Phi(X, M, J, K, d, t)
}
\arguments{
\item{X}{Design matrix of n observations and d covariables (numeric - n*d)}

\item{M}{M minus 1 is the degree of the polinomial (integer - M > 0)}

\item{J}{Númber of intervals in each dimention (integer - J > 1)}

\item{K}{Order of continuity in the derivatives (integrer - K < M)}

\item{d}{Number of dimentions, this parameter helps to improve efficiency}

\item{t}{matrix containing the nodes in which to split the Piecewise Polinomials}
}
\value{
A list of PWP expansion matrixes for each dimention d.
}
\description{
Calculates and returns a list of matrixes, each one representing a PWP
expansion for dimention d. Combines all of the parameters on a relatively
fast computation of basis expansion for X described on the thesis and on
(Denison, Mallik and Smith, 1998).
}
