% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{calculate_projection}
\alias{calculate_projection}
\title{Calculate Projection Vector}
\usage{
calculate_projection(F_mat, betas)
}
\arguments{
\item{F_mat}{The PWP transformed input space, calculated by a set of \code{w}
and a \code{Phi} matrix by the function \code{\link{calculate_F}}}

\item{betas}{The posterior puntual estimation for beta parameters calculated
by \code{\link{posterior_params}}}
}
\value{
A numeric vector representing the projection of \code{R^d} into
\code{R} given all of the parametres
}
\description{
Calculate Projection Vector
}
