% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{log_loss}
\alias{log_loss}
\title{Log Loss}
\usage{
log_loss(Y, p, eps = 1e-15, verb = TRUE)
}
\arguments{
\item{Y}{Response vector of n binary observatios (integers 0,1 - vector of
size n) Can be encoded as a factor a numeric vector.}

\item{p}{Vector of fitted probabilities for each case}

\item{eps}{Machine error to hanlde limit cases on the logarithm function}

\item{verb}{verbose, if True, prints aditional information (logical)}
}
\value{
The value of the Log Loss function (numeric). The smaller, the better
}
\description{
An implementation of the Log-Loss function for the binomial case
}
\examples{
log_loss(true_values, fitted probabilities)
log_loss(true_values, fitted probabilities)
#' log_loss(true_values, fitted probabilities, 1e-30, FALSE)
}
