% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{mode}
\alias{mode}
\title{Mode}
\usage{
mode(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The mode of the vector
}
\description{
Calculates the mode of a vector. Ties are resolved by the first element
}
