% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{model_projection}
\alias{model_projection}
\title{Calculates trained model for new Data}
\usage{
model_projection(new_X, bpwpm_params)
}
\arguments{
\item{new_X}{A new set of data for which to calculate the f(x) projection
function}

\item{bpwpm_params}{A list of bpwpm parameters created by the function
\code{\link{posterior_params}}}
}
\value{
A vector of the projection vector for a given set of points
}
\description{
Given a set of parameters inherited by the \code{\link{bpwpm_gibbs}} training
procedure, we calculate the value of the projection function for new data.
This function is both used on the predict and plot_3D functions.
}
\examples{
(test_data, mean_params)
}
