% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plot.bpwpm}
\alias{plot.bpwpm}
\title{Generic bpwpm plotting}
\usage{
\method{plot}{bpwpm}(object, n = 100, ...)
}
\arguments{
\item{object}{of the class bpwpm}

\item{n}{number of draws to plot}

\item{...}{additional parameters to be passed to the functions}
}
\value{
a series of line plots and histograms
}
\description{
Once a bpwpm has been run using the function \code{\link{bpwpm_gibbs}}, the
chains can be plotted and hence evaluated. This generic function builds sets
of plots for each parameter of the model, \eqn{\beta}, \eqn{w_1}, \eqn{w_2},
etc.
}
\examples{
(model1, 1000), (model2, 2000)
}
