% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plot_2D}
\alias{plot_2D}
\title{Wrapper Function for 2D Input Plots}
\usage{
plot_2D(Y, X, bpwpm_params, n, alpha = 0.6, f_of_0 = TRUE)
}
\arguments{
\item{Y}{A response vector of size n}

\item{X}{An input matrix of size n*2.}

\item{bpwpm_params}{an object of the class bpwpm_params of bpwpm_prediction
created by the functions \code{\link{posterior_params}} or
\code{\link{predict.bpwpm}} respectively that contains all the info about the
posterior parametres of the model.}

\item{n}{Thinness of grid for 2D and 3D projectction}

\item{alpha}{numeric - level of transparency for 2D projection}

\item{f_of_0}{Logical if If the constant function 0 is to be plotted}
}
\value{
A series of 3 plots to help ilustrate the model
}
\description{
To better understand the model, we can visualize it, however due to universal
limitations, plots are only availabe for X inputs of 2 dimentions. ie: only
two factors included on the regresion.
}
