% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_2D_proj}
\alias{plot_2D_proj}
\title{Plot 2D projection of the Model}
\usage{
plot_2D_proj(Y, X, bpwpm_params, n, alpha = 0.6)
}
\arguments{
\item{X}{Input Matrix of 2D (2 columns).}

\item{bpwpm_params}{A list of bpwpm parameters created by the function
\code{\link{posterior_params}}}

\item{n}{How thin is the grid to be made}

\item{alpha}{the corresponding alpha transparency param for the output space.}
}
\value{
A ggplot2 scatter plot
}
\description{
Once a model has been run and evaluated, in case that we have a 2D input
matrix, we can plot the projection to evaluate the model and its
corresponding binary outcomes. Instead of plotting the corresponding conotur
of the 3D function ploted by \code{\link{plot_3D_proj}} the output is
converted to its corresponding output and mapped to the 2D input space.
}
