% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plot_ergodic_mean}
\alias{plot_ergodic_mean}
\title{Plot Ergodic Mean}
\usage{
plot_ergodic_mean(object, thin, burn_in, ...)
}
\arguments{
\item{object}{of the class bpwpm}

\item{thin}{A thinning parameter for the MCMC Chain}

\item{burn_in}{A burn in parameter for the MCMC Chain}

\item{...}{additional parameters to be passed to the functions}
}
\value{
A series of plots for the ergodic mean of the parameters
}
\description{
Plots the Ergodic Mean of an object of class \code{bpwpm}
}
\examples{
(model1, 0, 0)
}
