% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{posterior_params}
\alias{posterior_params}
\title{Puntual estimation for parameters}
\usage{
posterior_params(bpwpm, thin, burn_in, type = "mean")
}
\arguments{
\item{thin}{A thinning parameter for the MCMC Chain}

\item{burn_in}{A burn in parameter for the MCMC Chain}

\item{type}{The type of punctual estimation for the parameters. Options
include: mean, mode or median.}
}
\value{
A parameter list that contains the estimations of beta and w.
}
\description{
Given a model output by the function bpwpm_gibbs, it thins down the chain and
makes puntual estimation for the parameters. This parameter structure can
later be used to calculate the F matrix and other results.
}
\examples{
(model, 2, 2000, 'mean') (model, 0, 0, 'median')
}
