% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_func.R
\name{simple_func}
\alias{simple_func}
\title{Simple Test Function}
\usage{
simple_func(x)
}
\arguments{
\item{x}{}
}
\value{
Returns a number squared
}
\description{
Squares an integer, updationg doc
}
\examples{
simple_func(2)
}
