% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpwpm_gibbs.R
\name{bpwpm_gibbs}
\alias{bpwpm_gibbs}
\title{Bayesian Piecewise Polinomial Model}
\usage{
bpwpm_gibbs(Y, X, M, J, K, intercept = TRUE, precision_beta = 1,
  precision_w = 1, draws = 10^3, z_init = NULL, beta_init = NULL,
  mu_beta_0 = NULL, sigma_beta_0_inv = NULL, w_init = NULL,
  mu_w_0 = NULL, sigma_w_0_inv = NULL, verb = TRUE, debug = FALSE,
  keep = TRUE)
}
\arguments{
\item{Y}{Response vector of n binary observatios (integers 0,1 - vector of
size n)}

\item{X}{Design matrix of n observations and d covariables (numeric - n*d)}

\item{M}{M minus 1 is the degree of the polinomial (integer - M > 0)}

\item{J}{Númber of intervals in each dimention (integer - J > 1)}

\item{K}{Order of continuity in the derivatives (integrer - K < M)}

\item{intercept}{Intercept:= To add an itercept term to the estimation. It's
recomended to have one (logical)}

\item{precision_beta}{If using the default sigmas for betas, a diagonal
matrix will be used. Precision controls its magnitude (numeric - precision
> 0)}

\item{precision_w}{If using the default sigmas for w, a diagonal matrix will
be used. Precision controls its magnitude (numeric - precision > 0)}

\item{draws}{Númber of samples to draw from the Gibbs Sampler (integer - draw
> 0)}

\item{z_init}{z_init:= Initial values for Auxiliary variable z (numeric -
vector of size n)}

\item{beta_init}{Initial value for the Gibbs Sampler Chain (numeric - vector
of size d)}

\item{mu_beta_0}{Prior Mean of Beta (numeric - vector of size d)}

\item{sigma_beta_0_inv}{Prior Inverse of the Variance-Covariance Matrix of
beta (numeric - matrix of size d*d)}

\item{w_init}{Inital value for the Gibbs Sampler Chain (numeric - matrix of
size N*d)}

\item{mu_w_0}{Prior Mean of Beta (numeric - matrix of size N*d)}

\item{sigma_w_0_inv}{sigma_w_0_inv:= Prior Inverse if the Variance-Covariance
Matrices of w (list - d elements, each element is a numeric matrix of size
N*N)}

\item{verb}{verbose, if True, prints aditional information (logical)}

\item{debug}{If TRUE, print even more info to help with debugging (logical)}

\item{keep}{If TRUE, saves a copy of the transformations Phi (logical)}
}
\value{
A list containing all the elements of the model
}
\description{
Gibbs Sampler for simulating draws of parameters for the Bayesian Piece Wise
Polinomial model described on the thesis.
}
\examples{
See the main document of the thesis for a couple of full examples
with its corresponding analysis.
}
