% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_2D_data}
\alias{plot_2D_data}
\title{Scatter Plot of 2D data}
\usage{
plot_2D_data(Y, X)
}
\arguments{
\item{Y}{A vector of binary response. Can be encoded as either a factor
vector or as a numeric one.}

\item{X}{Input Matrix of 2D (2 columns).}
}
\value{
A scatter plot of the 2 groups
}
\description{
Scatter Plot of 2D data
}
\examples{
(Y = rbinom(100, 1, 4), X = cbind(rnorm(100), rnorm(100)))
}
