% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plot_2D_proj}
\alias{plot_2D_proj}
\title{Plot 2D projection of the Model}
\usage{
plot_2D_proj(Y, X, bpwpm_params, n, alpha = 0.6)
}
\arguments{
\item{Y}{A response vector of size n}

\item{X}{An input matrix of size n*2.}

\item{bpwpm_params}{an object of the class bpwpm_params of bpwpm_prediction
created by the functions \code{\link{posterior_params}} or
\code{\link{predict.bpwpm}} respectively that contains all the info about the
posterior parametres of the model.}

\item{n}{Thinness of grid for 2D and 3D projectction}

\item{alpha}{numeric - level of transparency for 2D projection}
}
\value{
A ggplot2 scatter plot
}
\description{
2D projection of both the inputs and the posterior regions of classification.
Usefull to evaluate the corresponding binary outcomes.
Instead of plotting the corresponding conotur of the 3D function plotted by
\code{\link{plot_3D_proj}}. The projection function is mapped to it's
corresponding binary output and plotted behind the regular data.
}
