% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plot_chains}
\alias{plot_chains}
\title{Plot MCMC Chains}
\usage{
plot_chains(mcmc_chain, number = 100, title = "")
}
\arguments{
\item{mcmc_chain}{An MCMC Chain matrix. (draws * number of params)}

\item{number}{The number of draws to plot.}

\item{title}{Title for the plot}
}
\value{
A ggplot2 plot
}
\description{
Plots the last n draws of an MCMC chain
}
\examples{
plot_chains(betas), plot_chains(w_j, 1000)
}
